response

+++ volumeconverter.py
@@ -1,4 +1,4 @@
 print("How many gallons would you like to convert to liters?")
 gallons = input()
-liters = int(gallons) * 3.78541
+liters = float(gallons) * 3.78541
 print(str(gallons) + " gallons is " + str(liters) + " liters")
